+++ test/default/feature_test.go
@@ -1,6 +1,4 @@
-// Copyright (c) 2020-2022 Doc.ai and/or its affiliates.
-//
-// Copyright (c) 2023-2024 Cisco and/or its affiliates.
+// Copyright (c) 2024 Cisco and/or its affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -16,25 +14,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main_test
+package default_test
 
 import (
 	"testing"
 
 	"github.com/networkservicemesh/integration-tests/extensions/parallel"
-	"github.com/networkservicemesh/integration-tests/suites/basic"
-	"github.com/networkservicemesh/integration-tests/suites/ipsec_mechanism"
-	"github.com/networkservicemesh/integration-tests/suites/memory"
+	"github.com/networkservicemesh/integration-tests/suites/features"
 )
 
-func TestBasic(t *testing.T) {
-	parallel.Run(t, new(basic.Suite))
-}
-
-func TestMemory(t *testing.T) {
-	parallel.Run(t, new(memory.Suite))
-}
-
-func TestIPSec(t *testing.T) {
-	parallel.Run(t, new(ipsec_mechanism.Suite))
+func TestFeatureSuite(t *testing.T) {
+	parallel.Run(t, new(features.Suite), "TestVl3_ipv6", "TestVl3_dns", "TestVl3_lb", "TestVl3_scale_from_zero", "TestScale_from_zero", "TestSelect_forwarder")
 }
